Fix import and initial bugs, doesn't run
Fix the imports and some bugs. Now I have one other because in the
mutate I use the individual as a tour where in fact is already a
solution (therefore a np.array) and I have to fix it and check if it's
all correct just removeing the solution from every call in the method or
if I have to change more things.